Moves AdvocateCard to advocates folder
